+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ### Master
 
  * Fix ActiveModelSerializer polymorphic type processing for namespaced model names. Now AMS accepts '::' instead of '/'.
+ * `RESTSerializer#typeForRoot` has been deprecated. You can use
+`RESTSerializeer#modelNameFromPayloadKey` instead.
+ * Added `RESTSerializer#payloadKeyFromModelName`. This allows you to
+specify the outgoing root key for a JSON payload.
 
 ### Release 1.0.0-beta.16.1 (March 24, 2015)